+++ src/pds/registrysweepers/utils/db/host.py
@@ -1,0 +1,3 @@
+import collections
+
+Host = collections.namedtuple("Host", ["password", "url", "username", "verify"])
